+++ src/components/TestRail/Result.js
@@ -1,16 +1,19 @@
 class Result {
+
     /**
      *
      * @param caseId
      * @param statusId
      * @param comment
      * @param duration
+     * @param screenshotPath
      */
-    constructor(caseId, statusId, comment, duration) {
+    constructor(caseId, statusId, comment, duration, screenshotPath) {
         this._caseId = caseId;
         this._statusId = statusId;
         this._comment = comment;
         this._durationMS = duration;
+        this._screenshotPath = screenshotPath;
     }
 
     /**
@@ -40,6 +43,14 @@ class Result {
      */
     getElapsed() {
         return Math.round(this._durationMS / 1000) + 's';
+    }
+
+    /**
+     *
+     * @returns {*}
+     */
+    getScreenshotPath() {
+        return this._screenshotPath;
     }
 }